Improved store organization
